Copied over everything from converter
Starting with the width converter, we set up our special case of upsizing as a separate entity, because we don't think we know how to do the general case yet. And, the general case won't help us because we're not connecting up to a "real" AXIS interface (the DAC FIFO in the ADRV9371-zc706 reference design from Analog Devices).